FIX: Avoid possible error by checking for file before attempting deletion in create_previews_using_im function
git-svn-id: http://svn.resourcespace.com/svn/rs/trunk@10430 c08608d7-6e46-0410-86ca-f2a6f1370df5

+++ include/image_processing.php
@@ -1554,7 +1554,10 @@ function create_previews_using_im($ref,$thumbonly=false,$extension="jpg",$previe
 						$wait=run_command($cmd, true);
 						resource_log(RESOURCE_LOG_APPEND_PREVIOUS,LOG_CODE_TRANSFORMED,'','','',$cmd . ":\n" . $wait);
 
-						unlink($path);
+						if(file_exists($path))
+							{
+							unlink($path);
+							}
 						$path=str_replace($extension,"jpg",$path);
 						}
 					}
